+++ src/main/java/com/donrobo/fpbg/blueprint/AssemblingMachine2.java
@@ -36,12 +36,12 @@ public class AssemblingMachine2 extends AbstractBuilding {
     }
 
     @Override
-    protected int getBlueprintXOffset() {
+    protected double getBlueprintXOffset() {
         return 1;
     }
 
     @Override
-    protected int getBlueprintYOffset() {
+    protected double getBlueprintYOffset() {
         return 1;
     }
