+++ src/main/java/com/donrobo/fpbg/blueprint/FastInserter.java
@@ -31,12 +31,12 @@ public class FastInserter extends AbstractBuilding {
     }
 
     @Override
-    protected int getBlueprintXOffset() {
+    protected double getBlueprintXOffset() {
         return 0;
     }
 
     @Override
-    protected int getBlueprintYOffset() {
+    protected double getBlueprintYOffset() {
         return 0;
     }
